refactor(excel_export): move most function to lib
